// src/app/models/viabRepresentante.model .ts
export class viabRepresentanteModel {
    
    reprId: String;
    viabConsecutivo: String;
    viabNumeroIdentificacion: String;
    tipoIdentificacion: String;
    reprIdentificacion: string;
    reprNombre:String;
    cargoId: Number;
    reprEnRiesgo: Boolean;
    reprExisteEpicor: Boolean;
    reprComentarios: String;
   
    }